fix: multiple fight generation and standings improvements
Frontend:
- Add FightGenerationLoader component for better UX during fight generation
- Fix finishingMove display to correctly show winner's finishing move
- Fix genAIDescription property name bug in fight data
- Add comprehensive console logging for debugging fight data
- Add GET_ROUND_STANDINGS_BY_ROUND to refetchQueries for auto-update
- Add standings query state logging in DivisionPage

Backend:
- Fix genAIDescription property access (was fightDescription)
- Add field resolvers for Fighter objects in Fight type
- Fix competition type detection in validateFight function
- Change fight stats types from Int to Float for AI-generated decimals
- Remove strict significantStrikes validation that caused AI failures
- Add comprehensive logging for OpenAI responses and DB operations
- Fix fightId parameter to use ObjectId instead of fightIdentifier
- Implement smart standings resolver that returns initial/previous standings when none exist
- Add timing logs for OpenAI API and database operations

+++ server/inputs/fight-stats.input.js
@@ -17,9 +17,9 @@ const fightStatsInput = `#graphql
     input SignificantStrikesInput {
         accuracy: Float
         attempted: Int
-        defence: Int
+        defence: Float
         landed: Int
-        landedPerMinute: Int
+        landedPerMinute: Float
         positions: SignificantStrikePositionsInput
     }
 
@@ -41,15 +41,15 @@ const fightStatsInput = `#graphql
     }
 
     input SubmissionInput {
-        attemptsPer15Mins: Int
-        average: Int
+        attemptsPer15Mins: Float
+        average: Float
     }
 
     input TakedownInput {
         accuracy: Float
         attempted: Int
         avgTakedownsLandedPerMin: Float
-        defence: Int
+        defence: Float
         landed: Int
     }
 `;
